+++ postgres/src/main/java/org/jdbi/v3/postgres/PostgresPlugin.java
@@ -13,6 +13,7 @@
  */
 package org.jdbi.v3.postgres;
 
+import java.sql.Connection;
 import java.util.Map;
 import java.util.UUID;
 
@@ -49,6 +50,7 @@ import org.postgresql.util.PGmoney;
  * <li>{@link java.time.Period} (see notes below)</li>
  * <li>{@link java.util.Map Map&lt;String, String&gt;} (for {@code HSTORE} columns)</li>
  * <li>{@link java.util.UUID}</li>
+ * <li>{@link java.io.InputStream} and {@link java.io.Reader} from {@code oid} large object columns</li>
  * </ul>
  *
  * <p>
@@ -108,6 +110,8 @@ public class PostgresPlugin extends JdbiPlugin.Singleton {
         jdbi.registerArgument(new UUIDArgumentFactory());
         jdbi.registerArgument(new PGobjectArgumentFactory());
         jdbi.registerArgument(new BitStringEnumSetArgumentFactory());
+        jdbi.registerArgument(new BlobInputStreamArgumentFactory());
+        jdbi.registerArgument(new ClobReaderArgumentFactory());
 
         jdbi.registerArrayType(int.class, "integer");
         jdbi.registerArrayType(Integer.class, "integer");
@@ -138,6 +142,8 @@ public class PostgresPlugin extends JdbiPlugin.Singleton {
         jdbi.registerColumnMapper(new PeriodColumnMapperFactory());
         jdbi.registerColumnMapper(new PGobjectColumnMapperFactory());
         jdbi.registerColumnMapper(new BitStringEnumSetMapperFactory());
+        jdbi.registerColumnMapper(new BlobInputStreamColumnMapperFactory());
+        jdbi.registerColumnMapper(new ClobReaderColumnMapperFactory());
 
         // legacy unqualified HSTORE
         jdbi.registerArgument((ArgumentFactory) new HStoreArgumentFactory()::build);
@@ -150,8 +156,13 @@ public class PostgresPlugin extends JdbiPlugin.Singleton {
     }
 
     @Override
+    @SuppressWarnings("PMD.CloseResource")
     public Handle customizeHandle(Handle handle) {
-        PGConnection pgConnection = Unchecked.supplier(() -> handle.getConnection().unwrap(PGConnection.class)).get();
-        return handle.configure(PostgresTypes.class, pt -> pt.addTypesToConnection(pgConnection));
+        Connection conn = handle.getConnection();
+        PGConnection pgConnection = Unchecked.supplier(() -> conn.unwrap(PGConnection.class)).get();
+        return handle.configure(PostgresTypes.class, pt -> {
+            pt.addTypesToConnection(pgConnection);
+            pt.setLobApi(new PgLobApiImpl(conn));
+        });
     }
 }