+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # 3.14.0
   - immutables: support getter with @ColumnName, #1704
+  - postgres: simple CRUD support for LargeObject API
 
 # 3.13.0
   - Kotlin: respect default values in methods when passed null, #1690